# openapi.yaml
openapi: 3.0.3
info:
  title: LangGraph Multi-Agent System API
  description: |
    A comprehensive multi-agent workflow system built with FastAPI and LangGraph.

    This API provides endpoints for managing tasks, agents, workflows, and system operations.
    The system supports multiple execution modes, real-time communication via WebSockets,
    and comprehensive monitoring and health checking capabilities.

    ## Features
    - Task management with priority and type classification
    - Multi-agent coordination and execution
    - Real-time status updates via WebSocket
    - Comprehensive system monitoring and health checks
    - LLM configuration management
    - Batch operations and statistics
    - Meta-agent conversation management

    ## Authentication
    The API uses JWT-based authentication for securing endpoints.

    ## Rate Limiting
    Rate limiting is applied to prevent abuse and ensure system stability.

    ## WebSocket Support
    Real-time communication is supported through WebSocket endpoints for live updates.
  version: 1.0.0
  contact:
    name: LangGraph Multi-Agent Team
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/agents/{agent_id}/config:
    put:
      tags:
        - Agents
      summary: Update agent configuration
      description: Update configuration parameters for a specific agent
      operationId: updateAgentConfig
      parameters:
        - name: agent_id
          in: path
          required: true
          description: Agent ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                capabilities:
                  type: array
                  items:
                    type: string
                parameters:
                  type: object
                enabled:
                  type: boolean
      responses:
        '200':
          description: Agent configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/workflows/{workflow_id}/config:
    put:
      tags:
        - Workflows
      summary: Update workflow configuration
      description: Update configuration parameters for a specific workflow
      operationId: updateWorkflowConfig
      parameters:
        - name: workflow_id
          in: path
          required: true
          description: Workflow ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowConfigRequest'
      responses:
        '200':
          description: Workflow configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/llm/configs:
    get:
      tags:
        - LLM Configuration
      summary: List LLM configurations
      description: Retrieve a list of all LLM configurations
      operationId: listLLMConfigs
      responses:
        '200':
          description: LLM configurations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  configs:
                    type: array
                    items:
                      type: object
                      properties:
                        config_id:
                          type: string
                        name:
                          type: string
                        provider:
                          type: string
                        model:
                          type: string
                        active:
                          type: boolean
                        created_at:
                          type: string
                          format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - LLM Configuration
      summary: Create LLM configuration
      description: Create a new LLM configuration
      operationId: createLLMConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - provider
                - model
                - api_key
              properties:
                name:
                  type: string
                  description: Configuration name
                provider:
                  type: string
                  description: LLM provider
                  enum: [openai, anthropic, huggingface, local]
                model:
                  type: string
                  description: Model name
                api_key:
                  type: string
                  description: API key
                api_base:
                  type: string
                  description: API base URL
                parameters:
                  type: object
                  description: Model parameters
                description:
                  type: string
                  description: Configuration description
      responses:
        '201':
          description: LLM configuration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/llm/configs/{config_id}:
    get:
      tags:
        - LLM Configuration
      summary: Get LLM configuration
      description: Retrieve details of a specific LLM configuration
      operationId: getLLMConfig
      parameters:
        - name: config_id
          in: path
          required: true
          description: Configuration ID
          schema:
            type: string
      responses:
        '200':
          description: LLM configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  config_id:
                    type: string
                  name:
                    type: string
                  provider:
                    type: string
                  model:
                    type: string
                  api_base:
                    type: string
                  parameters:
                    type: object
                  description:
                    type: string
                  active:
                    type: boolean
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - LLM Configuration
      summary: Update LLM configuration
      description: Update an existing LLM configuration
      operationId: updateLLMConfig
      parameters:
        - name: config_id
          in: path
          required: true
          description: Configuration ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                provider:
                  type: string
                model:
                  type: string
                api_key:
                  type: string
                api_base:
                  type: string
                parameters:
                  type: object
                description:
                  type: string
      responses:
        '200':
          description: LLM configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - LLM Configuration
      summary: Delete LLM configuration
      description: Delete an LLM configuration
      operationId: deleteLLMConfig
      parameters:
        - name: config_id
          in: path
          required: true
          description: Configuration ID
          schema:
            type: string
      responses:
        '200':
          description: LLM configuration deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/llm/configs/{config_id}/test:
    post:
      tags:
        - LLM Configuration
      summary: Test LLM configuration
      description: Test an LLM configuration to verify connectivity and functionality
      operationId: testLLMConfig
      parameters:
        - name: config_id
          in: path
          required: true
          description: Configuration ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                test_prompt:
                  type: string
                  default: "Hello, please respond with a simple greeting."
                max_tokens:
                  type: integer
                  default: 50
      responses:
        '200':
          description: LLM configuration test completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  response_time_ms:
                    type: number
                  response:
                    type: string
                  tokens_used:
                    type: integer
                  error:
                    type: string
                    nullable: true
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/llm/configs/{config_id}/activate:
    post:
      tags:
        - LLM Configuration
      summary: Activate LLM configuration
      description: Set an LLM configuration as active for use
      operationId: activateLLMConfig
      parameters:
        - name: config_id
          in: path
          required: true
          description: Configuration ID
          schema:
            type: string
      responses:
        '200':
          description: LLM configuration activated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/llm/configs/{config_id}/usage:
    get:
      tags:
        - LLM Configuration
      summary: Get LLM usage statistics
      description: Retrieve usage statistics for a specific LLM configuration
      operationId: getLLMUsage
      parameters:
        - name: config_id
          in: path
          required: true
          description: Configuration ID
          schema:
            type: string
        - name: start_date
          in: query
          description: Start date for statistics
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: End date for statistics
          schema:
            type: string
            format: date
      responses:
        '200':
          description: LLM usage statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  config_id:
                    type: string
                  total_requests:
                    type: integer
                  total_tokens:
                    type: integer
                  total_cost:
                    type: number
                  average_response_time:
                    type: number
                  success_rate:
                    type: number
                  daily_usage:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          format: date
                        requests:
                          type: integer
                        tokens:
                          type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/llm/configs/{config_id}/usage/update:
    post:
      tags:
        - LLM Configuration
      summary: Update LLM usage data
      description: Update usage statistics for a specific LLM configuration
      operationId: updateLLMUsage
      parameters:
        - name: config_id
          in: path
          required: true
          description: Configuration ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tokens_used
                - response_time_ms
              properties:
                tokens_used:
                  type: integer
                  description: Number of tokens used
                response_time_ms:
                  type: number
                  description: Response time in milliseconds
                success:
                  type: boolean
                  description: Whether the request was successful
                cost:
                  type: number
                  description: Cost of the request
      responses:
        '200':
          description: LLM usage data updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    WorkflowConfigRequest:
      type: object
      properties:
        execution_mode:
          $ref: '#/components/schemas/WorkflowExecutionMode'
        checkpoint_interval:
          type: integer
          minimum: 10
          maximum: 3600
          description: Checkpoint interval in seconds
        max_retries:
          type: integer
          minimum: 0
          maximum: 10
          description: Maximum retry attempts
        timeout_seconds:
          type: integer
          minimum: 60
          maximum: 86400
          description: Timeout in seconds
        agents_config:
          type: object
          additionalProperties: true
          description: Agent configuration
...
    WorkflowExecutionMode:
      type: string
      enum: [sequential, parallel, adaptive]
      description: Workflow execution modes
...
    ApiResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: Operation success status
        message:
          type: string
          description: Response message
        data:
          type: object
          additionalProperties: true
          nullable: true
          description: Response data
        error_code:
          type: string
          nullable: true
          description: Error code if applicable
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      type: object
      required:
        - success
        - error_code
        - error_message
      properties:
        success:
          type: boolean
          default: false
          description: Always false for error responses
        error_code:
          type: string
          description: Machine-readable error code
        error_message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          nullable: true
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: BAD_REQUEST
            error_message: Invalid request parameters
            timestamp: "2024-01-01T00:00:00Z"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: UNAUTHORIZED
            error_message: Authentication required
            timestamp: "2024-01-01T00:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: NOT_FOUND
            error_message: Resource not found
            timestamp: "2024-01-01T00:00:00Z"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: INTERNAL_ERROR
            error_message: Internal server error
            timestamp: "2024-01-01T00:00:00Z"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
security:
  - BearerAuth: []
...